Rearranging Files and Updating Briefings

+++ src/main/java/br/com/grupobrasileiro/sistema_grupo_brasileiro_backend/dto/user/view/EmployeeView.java
@@ -57,6 +57,6 @@ public record EmployeeView(
          */
         String agency,
 
-        Integer avatar
+        Long avatar
 ) {
 }
